12_OpenAI_Agent_SDK/08_Agent_MCP/pyproject.toml
[project]
name = "08-agent-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.10.1",
    "mcp>=1.10.1",
    "openai-agents>=0.1.0",
    "pydantic-ai>=0.3.5",
    "python-dotenv>=1.1.1",
]
